added shared module (steppable) for testing pdes
